djangofest/requirements.txt
amqp==2.6.1
argh==0.26.2
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.0
importlib-metadata==2.0.0
kombu==4.6.11
pathtools==0.1.2
psycopg2-binary==2.8.6
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
watchdog==0.10.3
zipp==3.2.0
djangorestframework==3.11
drf-yasg==1.17.1
django-cors-headers==3.5.0
django-3-jet==1.0.8
pillow
django-prettyjson